requirements.txt
Flask
scikit-learn
pdfplumber
reportlab
python-dotenv
requests
google-generativeai
werkzeug
sentence-transformers
spacy
